Reorganize specs module
